+++ src/Sentry.Unity.Editor/SentryWindow.cs
@@ -262,7 +262,8 @@ namespace Sentry.Unity.Editor
 
             Options.InitCacheFlushTimeout = EditorGUILayout.IntField(
                 new GUIContent("Init Flush Timeout [ms]", "The timeout that limits how long the SDK " +
-                                                          "will attempt to flush existing cache during initialization." +
+                                                          "will attempt to flush existing cache during initialization, " +
+                                                          "potentially slowing down app start up to the specified time." +
                                                           "\nThis features allows capturing errors that happen during " +
                                                           "game startup and would not be captured because the process " +
                                                           "would be killed before Sentry had a chance to capture the event."),